backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
python-dotenv==1.0.1
numpy==1.26.4
scikit-fuzzy==0.4.2 